+++ packages/design-system-web/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.4.4-pr-27.2 (2022-10-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.4.4-pr-27.1 (2022-10-12)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.4.4-pr-27.0 (2022-10-12)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
 ## 1.4.3 (2022-10-06)
 
 **Note:** Version bump only for package @geovistory/design-system-web